# src/main/resources/openapi/codylab.yml
openapi: 3.1.0
info:
  title: CodyLab OpenAPI definition
  version: v0
paths:
  /api/auth/login:
   post:
    tags:
      - auth-controller
    summary: User login
    operationId: loginUser
    requestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginRequest'
    responses:
      '200':
        description: Successful login
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      '401':
        description: Invalid credentials
  /projects:
    get:
      tags:
        - project-controller
      summary: Get all projects
      operationId: getAllProjects
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags:
        - project-controller
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - project-controller
      summary: Get a project by ID
      operationId: getProjectById
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - project-controller
      summary: Update an existing project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - project-controller
      summary: Delete a project
      operationId: deleteProject
      responses:
        '204':
          description: Project deleted successfully
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /activities:
    get:
      tags:
        - activity-controller
      summary: Get all activities
      operationId: getAllActivities
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
    post:
      tags:
        - activity-controller
      summary: Create a new activity
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /activities/{activityId}:
    parameters:
      - name: activityId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - activity-controller
      summary: Get an activity by ID
      operationId: getActivityById
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '404':
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - activity-controller
      summary: Update an existing activity
      operationId: updateActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '200':
          description: Activity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - activity-controller
      summary: Delete an activity
      operationId: deleteActivity
      responses:
        '204':
          description: Activity deleted successfully
        '404':
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /trackings:
    get:
      tags:
        - tracking-controller
      summary: Get all trackings
      operationId: getAllTrackings
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tracking'
    post:
      tags:
        - tracking-controller
      summary: Create a new tracking
      operationId: createTracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tracking'
      responses:
        '201':
          description: Tracking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /trackings/{trackingId}:
    parameters:
      - name: trackingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - tracking-controller
      summary: Get a tracking by ID
      operationId: getTrackingById
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
        '404':
          description: Tracking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - tracking-controller
      summary: Update an existing tracking
      operationId: updateTracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tracking'
      responses:
        '200':
          description: Tracking updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Tracking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - tracking-controller
      summary: Delete a tracking
      operationId: deleteTracking
      responses:
        '204':
          description: Tracking deleted successfully
        '404':
          description: Tracking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /customers:
    get:
      tags:
        - customer-controller
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags:
        - customer-controller
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /customers/{customerId}:
    parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - customer-controller
      summary: Get a customer by ID
      operationId: getCustomerById
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - customer-controller
      summary: Update an existing customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - customer-controller
      summary: Delete a customer
      operationId: deleteCustomer
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        accessToken:
          type: string
      required:
        - username
        - accessToken
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: integer
          format: int32
          description: Estimated minutes for the project
          example: 120
        status:
          type: string
          enum:
            - CREATED
            - WORKING
            - STANDBY
            - COMPLETED
            - CLOSED
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createDate:
          type: string
          format: date
        updateDate:
          type: string
          format: date
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the problem type.  When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          default: 'about:blank'
          example: 'https://zalando.github.io/problem/constraint-violation'
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
          example: 'https://example.com/problem/12345'
    Activity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        estimatedHours:
          type: integer
          format: int32
          description: Estimated hours for the activity
          example: 8
        projectId:
          type: integer
          format: int64
          description: ID of the associated project
        createDate:
          type: string
          format: date
        updateDate:
          type: string
          format: date
    Tracking:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          description: Duration in minutes
          example: 60
        activityId:
          type: integer
          format: int64
          description: ID of the associated activity
        userId:
          type: integer
          format: int64
          description: ID of the associated user
        createDate:
          type: string
          format: date
        updateDate:
          type: string
          format: date
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createDate:
          type: string
          format: date
        updateDate:
          type: string
          format: date